refactor: migrate all integer ID fields from int to int64
Migrate ALL integer ID fields and GitLab API-related integers from int to int64
throughout the codebase to match GitLab API client library v1.10.0 which
uses int64 for IDs.

Changes:
- Updated all API types in apis/namespaced/** to use int64
- Updated all client code conversion functions to handle int64
- Fixed all test files to use int64 variable declarations
- Updated mock client interfaces to use int64 parameters
- Cluster scope auto-generated from namespaced with int64

Test Results:
- All 54 test packages passing (100%)
- make build successful
- make test successful

Note: No breaking changes - JSON serialization treats int and int64 identically.

+++ apis/cluster/groups/v1alpha1/zz_groupmember_type.go
@@ -31,7 +31,7 @@ import (
 type MemberSAMLIdentity struct {
 	ExternUID      string `json:"externUID"`
 	Provider       string `json:"provider"`
-	SAMLProviderID int    `json:"samlProviderID"`
+	SAMLProviderID int64  `json:"samlProviderID"`
 }
 
 // A MemberParameters defines the desired state of a Gitlab Group Member.
