+++ apis/namespaced/groups/v1alpha1/group_types.go
@@ -241,7 +241,7 @@ type SharedWithGroups struct {
 	// https://docs.gitlab.com/ee/api/members.html#roles
 	// +required
 	// +immutable
-	GroupAccessLevel int `json:"groupAccessLevel"`
+	GroupAccessLevel int64 `json:"groupAccessLevel"`
 
 	// Share expiration date in ISO 8601 format: 2016-09-26
 	// +optional
@@ -251,7 +251,7 @@ type SharedWithGroups struct {
 
 // GroupObservation is the observed state of a Group.
 type GroupObservation struct {
-	ID                  *int                          `json:"id,omitempty"`
+	ID                  *int64                        `json:"id,omitempty"`
 	AvatarURL           *string                       `json:"avatarUrl,omitempty"`
 	WebURL              *string                       `json:"webUrl,omitempty"`
 	FullName            *string                       `json:"fullName,omitempty"`
@@ -268,10 +268,10 @@ type GroupObservation struct {
 
 // SharedWithGroupsObservation is the observed state of a SharedWithGroups.
 type SharedWithGroupsObservation struct {
-	GroupID          *int         `json:"groupId,omitempty"`
+	GroupID          *int64       `json:"groupId,omitempty"`
 	GroupName        *string      `json:"groupName,omitempty"`
 	GroupFullPath    *string      `json:"groupFullPath,omitempty"`
-	GroupAccessLevel *int         `json:"groupAccessLevel,omitempty"`
+	GroupAccessLevel *int64       `json:"groupAccessLevel,omitempty"`
 	ExpiresAt        *metav1.Time `json:"expiresAt,omitempty"`
 }
